# Fogadoora Backend/src/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "REST API minta backend",
        "version": "1.0.0",
        "description": "<img alt='HaHu' height='400px' src='https://nitslaszlo.github.io/static/hahu/Hahu.jpg' />"
    },
    "servers": [
        {
            "url": "http://localhost:5000"
        },
        {
            "url": "https://jedlik.cyclic.app"
        }
    ],
    "paths": {
        "/api/reservations": {
            "get": {
                "tags": ["reservations"],
                "summary": "Összes foglalás lekérdezése.",
                "description": "Az összes foglalás adatát tudjuk lekérdezni.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/reservation"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "error.message"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["reservations"],
                "summary": "Új foglalás mentése.",
                "description": "Az új foglalas mentése a reservations táblában",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "$ref": "#/components/schemas/reservation"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK (Az új, elmentett foglalással tér vissza.)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/reservation"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "error.message"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/reservations/{offset}/{limit}/{field}": {
            "get": {
                "tags": ["reservations"],
                "summary": "Paginált adatok lekérése.",
                "description": "Foglalások lekérdezése lapozáshoz.",
                "parameters": [
                    {
                        "name": "offset",
                        "description": "Ennyi rekordot hagyunk el (skip) a lekérdezés 'elejéről'.",
                        "in": "path",
                        "schema": {
                            "type": "number",
                            "example": "5"
                        },
                        "required": true
                    },
                    {
                        "name": "limit",
                        "description": "Ennyi rekordot küld (take) maximum az API",
                        "in": "path",
                        "schema": {
                            "type": "number",
                            "example": "5"
                        },
                        "required": true
                    },
                    {
                        "name": "field",
                        "description": "Melyik mező szerint legyenek rendezve a rekordok. A kötőjel '-' karakter a csökkenő rendezést jelöli.",
                        "in": "path",
                        "schema": {
                            "type": "string",
                            "example": "-cim"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "count": {
                                            "type": "number",
                                            "description": "Az összes rekord száma.",
                                            "example": "16"
                                        },
                                        "data": {
                                            "type": "array",
                                            "description": "A maximum {limit} számú rekord.",
                                            "items": {
                                                "$ref": "#/components/schemas/reservation"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "error.message"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/reservations/{offset}/{limit}/{field}/{filter}": {
            "get": {
                "tags": ["reservations"],
                "summary": "Paginált adatok lekérése szűréssel.",
                "description": "Foglalások lekérdezése lapozáshoz szűréssel.",
                "parameters": [
                    {
                        "name": "offset",
                        "description": "Ennyi rekordot hagyunk el (skip) a lekérdezés 'elejéről'.",
                        "in": "path",
                        "schema": {
                            "type": "number",
                            "example": "0"
                        },
                        "required": true
                    },
                    {
                        "name": "limit",
                        "description": "Ennyi rekordot küld (take) maximum az API.",
                        "in": "path",
                        "schema": {
                            "type": "number",
                            "example": "2"
                        },
                        "required": true
                    },
                    {
                        "name": "field",
                        "description": "Melyik mező szerint legyenek rendezve a rekordok. A kötőjel '-' karakter a csökkenő rendezést jelöli.",
                        "in": "path",
                        "schema": {
                            "type": "string",
                            "example": "-cim"
                        },
                        "required": true
                    },
                    {
                        "name": "filter",
                        "description": "Szűrő - a 'cim' és a 'leiras' mezőkben keressük.",
                        "in": "path",
                        "schema": {
                            "type": "string",
                            "example": "sport"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "count": {
                                            "type": "number",
                                            "description": "A szűrőnek megfelelő rekordok száma",
                                            "example": "16"
                                        },
                                        "data": {
                                            "type": "array",
                                            "description": "A maximum {limit} számú, szűrőnek megfelelő rekord.",
                                            "items": {
                                                "$ref": "#/components/schemas/reservation"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "error.message"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "reservation": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "description": "PK - A foglalás azonosítója (ObjectID).",
                        "example": "61daf48ba5775848e99f555b"
                    },
                    "idő": {
                        "type": "string",
                        "description": "Az időpont, amikor a tanárral le lehet ülni beszélni.",
                        "example": "17:40"
                    },
                    "dátum": {
                        "type": "string",
                        "description": "A dátum, amikor rögzítve lett a foglalás.",
                        "example": "2023.03.01."
                    },
                    "tanárID": {
                        "type": "string",
                        "description": "A kiválasztott tanár azonosítója.",
                        "example": "8"
                    }
                }
            },
...
        }
    }
}